+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 0.24.0 - in development
 - Fix `thingy_to_region` name-row
 - Delay selection highlight
+- [Create .clj-kondo directory before analysis](https://github.com/pedrorgirardi/Pep/issues/57)
 
 ## 0.23.0 - 2023-12-24
 - Find symbol usages from symbol